fix(readLimit): fix error type

+++ client.go
@@ -20,7 +20,7 @@ var (
 	// Send pings to peer with this period. Must be less than pongWait.
 	pingPeriod = (pongWait * 9) / 10
 	// Maximum message size allowed from peer.
-	maxMessageSize = 0
+	maxMessageSize = int64(0)
 )
 
 type clientHandler struct {
